Save work on zustand approach
