+++ packages/studio-base/src/context/ExtensionRegistryContext.ts
@@ -2,9 +2,6 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import { createContext, useContext } from "react";
-import { DeepReadonly } from "ts-essentials";
-
 import { ExtensionPanelRegistration } from "@foxglove/studio";
 import { ExtensionInfo, ExtensionNamespace } from "@foxglove/studio-base/types/Extensions";
 
@@ -21,23 +18,8 @@ export type ExtensionRegistry = {
     foxeFileData: Uint8Array,
   ) => Promise<ExtensionInfo>;
   loadExtension(id: string): Promise<string>;
-  refreshExtensions: () => Promise<DeepReadonly<ExtensionInfo[]>>;
-  registeredExtensions: ExtensionInfo[];
-  registeredPanels: Record<string, RegisteredPanel>;
-  uninstallExtension(namespace: ExtensionNamespace, id: string): Promise<void>;
+  refreshExtensions: () => Promise<void>;
+  registeredExtensions: undefined | ExtensionInfo[];
+  registeredPanels: undefined | Record<string, RegisteredPanel>;
+  uninstallExtension: (namespace: ExtensionNamespace, id: string) => Promise<void>;
 };
-
-const ExtensionRegistryContext = createContext<DeepReadonly<ExtensionRegistry> | undefined>(
-  undefined,
-);
-ExtensionRegistryContext.displayName = "ExtensionRegistryContext";
-
-export function useExtensionRegistry(): DeepReadonly<ExtensionRegistry> {
-  const extensionRegistry = useContext(ExtensionRegistryContext);
-  if (extensionRegistry == undefined) {
-    throw new Error("An ExtensionRegistryContext provider is required to useExtensionRegistry");
-  }
-  return extensionRegistry;
-}
-
-export default ExtensionRegistryContext;
